src/requirements.txt
SQLAlchemy
PyYAML
ccxt
psycopg2